fixing bugs
Refactor bit field handling and add new utilities

Updated PackedStruct to use specific field names for BitField attributes and added a new private field. Modified type checks and assignments in BaseAttributeParser. Changed dictionary usage and corrected conditions in BaseBitFieldPropertyAggregator. Added Microsoft.Bcl.HashCode package reference. Introduced ExistingFieldAggregator in PropertyBitPackSourceGenerator. Improved exception message formatting in ThrowHelper. Added new classes for custom equality comparer and dictionary pool management.

+++ PropertyBitPack.SourceGen/AttributeParsers/BaseAttributeParser.cs
@@ -192,9 +192,9 @@ internal abstract class BaseAttributeParser : IAttributeParser, IContextBindable
         {
             validBitsCount = null;
         }
-        else if(candidateBitsCount is byte v)
+        else if(candidateBitsCount is int v)
         {
-            validBitsCount = v;
+            validBitsCount = (byte)v;
         }
         else
         {
